refactor: pull out ColumnDef
